+++ app/rust/src/frb_generated.rs
@@ -39,7 +39,7 @@ flutter_rust_bridge::frb_generated_boilerplate!(
     default_rust_auto_opaque = RustAutoOpaqueMoi,
 );
 pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_VERSION: &str = "2.12.0";
-pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = 2025840930;
+pub(crate) const FLUTTER_RUST_BRIDGE_CODEGEN_CONTENT_HASH: i32 = 1014381508;
 
 // Section: executor
 
@@ -1404,6 +1404,57 @@ fn wire__crate__api__umbra__UmbraApp_rename_group_impl(
         },
     )
 }
+fn wire__crate__api__umbra__UmbraApp_repair_tor_impl(
+    ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
+    rust_vec_len_: i32,
+    data_len_: i32,
+) -> flutter_rust_bridge::for_generated::WireSyncRust2DartSse {
+    FLUTTER_RUST_BRIDGE_HANDLER.wrap_sync::<flutter_rust_bridge::for_generated::SseCodec, _>(
+        flutter_rust_bridge::for_generated::TaskInfo {
+            debug_name: "UmbraApp_repair_tor",
+            port: None,
+            mode: flutter_rust_bridge::for_generated::FfiCallMode::Sync,
+        },
+        move || {
+            let message = unsafe {
+                flutter_rust_bridge::for_generated::Dart2RustMessageSse::from_wire(
+                    ptr_,
+                    rust_vec_len_,
+                    data_len_,
+                )
+            };
+            let mut deserializer =
+                flutter_rust_bridge::for_generated::SseDeserializer::new(message);
+            let api_that = <RustOpaqueMoi<
+                flutter_rust_bridge::for_generated::RustAutoOpaqueInner<UmbraApp>,
+            >>::sse_decode(&mut deserializer);
+            let api_sink = <StreamSink<
+                crate::api::umbra::NetEvent,
+                flutter_rust_bridge::for_generated::SseCodec,
+            >>::sse_decode(&mut deserializer);
+            deserializer.end();
+            transform_result_sse::<_, String>((move || {
+                let mut api_that_guard = None;
+                let decode_indices_ =
+                    flutter_rust_bridge::for_generated::lockable_compute_decode_order(vec![
+                        flutter_rust_bridge::for_generated::LockableOrderInfo::new(
+                            &api_that, 0, false,
+                        ),
+                    ]);
+                for i in decode_indices_ {
+                    match i {
+                        0 => api_that_guard = Some(api_that.lockable_decode_sync_ref()),
+                        _ => unreachable!(),
+                    }
+                }
+                let api_that_guard = api_that_guard.unwrap();
+                let output_ok =
+                    crate::api::umbra::UmbraApp::repair_tor(&*api_that_guard, api_sink)?;
+                Ok(output_ok)
+            })())
+        },
+    )
+}
 fn wire__crate__api__umbra__UmbraApp_search_messages_impl(
     ptr_: flutter_rust_bridge::for_generated::PlatformGeneralizedUint8ListPtr,
     rust_vec_len_: i32,
@@ -2561,10 +2612,10 @@ fn pde_ffi_dispatcher_primary_impl(
 ) {
     // Codec=Pde (Serialization + dispatch), see doc to use other codecs
     match func_id {
-        40 => {
+        41 => {
             wire__crate__api__umbra__UmbraApp_start_network_impl(port, ptr, rust_vec_len, data_len)
         }
-        45 => wire__crate__api__simple__init_app_impl(port, ptr, rust_vec_len, data_len),
+        46 => wire__crate__api__simple__init_app_impl(port, ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }
@@ -2615,28 +2666,29 @@ fn pde_ffi_dispatcher_sync_impl(
         28 => wire__crate__api__umbra__UmbraApp_pending_messages_impl(ptr, rust_vec_len, data_len),
         29 => wire__crate__api__umbra__UmbraApp_rename_contact_impl(ptr, rust_vec_len, data_len),
         30 => wire__crate__api__umbra__UmbraApp_rename_group_impl(ptr, rust_vec_len, data_len),
-        31 => wire__crate__api__umbra__UmbraApp_search_messages_impl(ptr, rust_vec_len, data_len),
-        32 => wire__crate__api__umbra__UmbraApp_send_file_impl(ptr, rust_vec_len, data_len),
-        33 => {
+        31 => wire__crate__api__umbra__UmbraApp_repair_tor_impl(ptr, rust_vec_len, data_len),
+        32 => wire__crate__api__umbra__UmbraApp_search_messages_impl(ptr, rust_vec_len, data_len),
+        33 => wire__crate__api__umbra__UmbraApp_send_file_impl(ptr, rust_vec_len, data_len),
+        34 => {
             wire__crate__api__umbra__UmbraApp_send_group_message_impl(ptr, rust_vec_len, data_len)
         }
-        34 => wire__crate__api__umbra__UmbraApp_send_over_network_impl(ptr, rust_vec_len, data_len),
-        35 => wire__crate__api__umbra__UmbraApp_set_autologin_impl(ptr, rust_vec_len, data_len),
-        36 => wire__crate__api__umbra__UmbraApp_set_contact_saved_impl(ptr, rust_vec_len, data_len),
-        37 => {
+        35 => wire__crate__api__umbra__UmbraApp_send_over_network_impl(ptr, rust_vec_len, data_len),
+        36 => wire__crate__api__umbra__UmbraApp_set_autologin_impl(ptr, rust_vec_len, data_len),
+        37 => wire__crate__api__umbra__UmbraApp_set_contact_saved_impl(ptr, rust_vec_len, data_len),
+        38 => {
             wire__crate__api__umbra__UmbraApp_set_contact_status_impl(ptr, rust_vec_len, data_len)
         }
-        38 => {
+        39 => {
             wire__crate__api__umbra__UmbraApp_set_custom_bridges_impl(ptr, rust_vec_len, data_len)
         }
-        39 => wire__crate__api__umbra__UmbraApp_set_my_picture_impl(ptr, rust_vec_len, data_len),
-        41 => wire__crate__api__umbra__UmbraApp_user_code_impl(ptr, rust_vec_len, data_len),
-        42 => wire__crate__api__umbra__UmbraApp_username_impl(ptr, rust_vec_len, data_len),
-        43 => wire__crate__api__umbra__app_version_impl(ptr, rust_vec_len, data_len),
-        44 => wire__crate__api__simple__greet_impl(ptr, rust_vec_len, data_len),
-        46 => wire__crate__api__umbra__install_update_impl(ptr, rust_vec_len, data_len),
-        47 => wire__crate__api__umbra__offered_update_impl(ptr, rust_vec_len, data_len),
-        48 => wire__crate__api__umbra__set_native_dir_impl(ptr, rust_vec_len, data_len),
+        40 => wire__crate__api__umbra__UmbraApp_set_my_picture_impl(ptr, rust_vec_len, data_len),
+        42 => wire__crate__api__umbra__UmbraApp_user_code_impl(ptr, rust_vec_len, data_len),
+        43 => wire__crate__api__umbra__UmbraApp_username_impl(ptr, rust_vec_len, data_len),
+        44 => wire__crate__api__umbra__app_version_impl(ptr, rust_vec_len, data_len),
+        45 => wire__crate__api__simple__greet_impl(ptr, rust_vec_len, data_len),
+        47 => wire__crate__api__umbra__install_update_impl(ptr, rust_vec_len, data_len),
+        48 => wire__crate__api__umbra__offered_update_impl(ptr, rust_vec_len, data_len),
+        49 => wire__crate__api__umbra__set_native_dir_impl(ptr, rust_vec_len, data_len),
         _ => unreachable!(),
     }
 }